Refactor token imports to new service package.
