+++ pkg/platforms/batocera/platform.go
@@ -3,6 +3,7 @@ package batocera
 import (
 	"errors"
 	"github.com/wizzomafizzo/tapto/pkg/api/models"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -16,7 +17,6 @@ import (
 	"github.com/wizzomafizzo/tapto/pkg/readers/file"
 	"github.com/wizzomafizzo/tapto/pkg/readers/libnfc"
 	"github.com/wizzomafizzo/tapto/pkg/readers/simple_serial"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
 )
 
 type Platform struct {
