+++ pkg/platforms/mistex/platform.go
@@ -5,6 +5,7 @@ package mistex
 import (
 	"fmt"
 	"github.com/wizzomafizzo/tapto/pkg/api/models"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -23,7 +24,6 @@ import (
 	"github.com/wizzomafizzo/tapto/pkg/readers/file"
 	"github.com/wizzomafizzo/tapto/pkg/readers/libnfc"
 	"github.com/wizzomafizzo/tapto/pkg/readers/simple_serial"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
 )
 
 type Platform struct {
