+++ pkg/platforms/platforms.go
@@ -3,12 +3,12 @@ package platforms
 import (
 	"github.com/wizzomafizzo/tapto/pkg/api/models"
 	"github.com/wizzomafizzo/tapto/pkg/service/playlists"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"path/filepath"
 	"strings"
 
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/readers"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
 )
 
 type CmdEnv struct {
