+++ pkg/readers/file/file.go
@@ -3,6 +3,7 @@ package file
 import (
 	"encoding/hex"
 	"errors"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"os"
 	"path/filepath"
 	"strings"
@@ -11,7 +12,6 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/readers"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
 	"github.com/wizzomafizzo/tapto/pkg/utils"
 )
