+++ pkg/readers/optical_drive/optical_drive.go
@@ -5,7 +5,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/readers"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"github.com/wizzomafizzo/tapto/pkg/utils"
 	"os"
 	"os/exec"
