+++ pkg/readers/simple_serial/simple_serial.go
@@ -2,6 +2,7 @@ package simple_serial
 
 import (
 	"errors"
+	"github.com/wizzomafizzo/tapto/pkg/service/tokens"
 	"os"
 	"runtime"
 	"strings"
@@ -10,7 +11,6 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/readers"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
 	"github.com/wizzomafizzo/tapto/pkg/utils"
 
 	"go.bug.st/serial"
